+++ Infrastructure/BuilderProviderProtocol.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public protocol BuilderProviderProtocol {
+
+    associatedtype ResponseType: Decodable
+    associatedtype ModelType
+
+    func build(response: ResponseType) throws -> ModelType
+}
