+++ Podfile
@@ -1,4 +1,4 @@
-platform :ios, '13.0'
+platform :ios, '14.0'
 
 def unitTestingPods
   pod 'Quick', '7.4.1'
@@ -10,13 +10,8 @@ def visualPods
   pod 'SDWebImage', '5.19.0'
 end
 
-def infrastructurePods
-  pod 'PromiseKit', '8.0.0'
-end
-
 target 'SwiftmazingMock' do
   use_frameworks!
-  infrastructurePods
   visualPods
 
   target 'SwiftmazingFunctionalTests' do
@@ -28,13 +23,11 @@ end
 
 target 'Swiftmazing' do
   use_frameworks!
-  infrastructurePods
   visualPods
 
   target 'SwiftmazingTests' do
     inherit! :search_paths
     unitTestingPods
-    infrastructurePods
     visualPods
   end
 
@@ -52,7 +45,6 @@ end
 
 target 'Infrastructure' do
   use_frameworks!
-  infrastructurePods
 
   target 'InfrastructureTests' do
     unitTestingPods
@@ -64,7 +56,7 @@ post_install do |installer_representation|
   installer_representation.pods_project.targets.each do |target|
       target.build_configurations.each do |config|
           config.build_settings['CLANG_ENABLE_CODE_COVERAGE'] = 'NO'
-          config.build_settings['IPHONEOS_DEPLOYMENT_TARGET'] = 13
+          config.build_settings['IPHONEOS_DEPLOYMENT_TARGET'] = 14
           
       end
   end
